Expose context to notice and notification handlers
Add new handler types for notices and notifications. These are
equivalent to the existing types, but receive a context as the first
parameter. This allows for notices and notifications to use values
typically stored in contexts, such as logging/tracing metadata.

The existing behavior is kept for backwards compatability, but the
`OnNotice` and `OnNotifiation` fields of `pgconn.Config` are
deprecated. a