Add more performant CollectRows functions
Add the functions CollectRowsUsing/Scanner et al. as more performant
alternatives to CollectRows et al.

The new functions have the same names as the old one, suffixed with
"Using" or "UsingScanner". The "UsingScanner" functions take a
RowScannerInto object as an argument that provides similar functionaly
to a RowScanner, but with some extensions and a generic type parameter.

The "Using" functions wrap the "UsingScanner" functions. They take a
function as an argument that generates a RowScannerInto object. These
functions have the same names as the "RowTo" functions, but with "RowTo"
replaced with "RowInto". These "Using" functions provide a more-similar
interface to the old functions, and guarantees about the lifetime of the
RowScannerInto object - this can enable further optimizations like object
pooling.

Ideally, the already-existing methods could have been modified, rather
than duplicated. I couldn't find a way to do this without changing the
type of RowToFunc, which would technically represent a breaking API
change. The next major-versioned release should probably consolidate
these methods.

Because CollectRows uses a RowToFunc to describe how to scan the row
(and because these functions are totally opaque to the caller and return
a value constructed internally) they inherently induce unnecessary
allocations and copies. Additionally, struct types require some work to
compute the mapping from columns to fields - this could be computed once
and reused between rows, but the interface provides no easy way to do
this and instead requires the work to be redone for each row.

The new interface allows the following optimizations:

* By passing a pointer receiver into ScanRowInto rather than returning
  the scanned result, we can serialize directly into the returned slice
  and avoid an additional allocation and copy for each value.
* By using a scanner object with state rather than just a function, we
  can reuse allocated memory for scannerTargets rather than reallocating
  for each row.
* By providing an Initialize method, we can pre-compute the mapping from
  column -> field for structs just once per query, rather than for each
  row.