Close batch pipeline after a query error
Currently, if a batch ends with a statement_timeout on a br.Query(),
the next br.Close() will only return the query error and leave the
underlying pipeline opened. This prevents the connection from being
reused and a second br.Close() is required to really close the pipeline.

This patch changes br.Close() to fallthrough and close the pipeline,
even with a lastRow error present.