Fix Rows.FieldDescriptions for empty query
Empty queries are now treated as results rather than no result.

This caused a problem for an empty batch in the simple protocol because
the batch request was sending an empty string which would now be
interpreted by the batch results as an empty query result. The solution
for this was to catch empty batches in SendBatch and return a
specialized empty batch results implementation. So we also save the
round trip to the server for empty batches.

https://github.com/jackc/pgx/issues/2402