Prepare overwrites existing prepared statement
Prepare now sends a Close message as part of preparing a statement. This
allows changes Prepare behavior in that it is no longer an error to
prepare a statement with a name that already exists. In addition, it
resolves issues where the connection gets confused and doesn't realize
a statement is already prepared. This could happen if a statement_timout
triggered between the Parse and Describe messages.

https://github.com/jackc/pgx/issues/2223