Handle for preparing statements that fail during the Describe phase
A failure that occurred after a successful Parse could leave a prepared
statement on the server that must be deallocated before re-preparing.

Also, introduce a testing utility faultyconn.Conn that can be used to
intercept and modify frontend messages. This is used to induce difficult
difficult to reproduce timing and protocol errors in tests.