Fix: json/jsonb sql.Scanner source type is []byte
Previously, when scanning a json/jsonb value into a sql.Scanner via
database/sql the source value was a []byte. However, when scanning into
a sql.Scanner via pgx the source value was a string. This change ensures
that the source value is consistently a []byte.

The problem was pgx has one scan plan for sql.Scanner when a Codec is
available and another when it is not. When a Codec is not available,
this means we don't know the underlying type, so all we can do is pass
text formatted values as a string and binary formatted values as a
[]byte. However, when a Codec is available we can use the Codec to
call DecodeDatabaseSQLValue which provides the proper source type.
JSONCodec.planScan was incorrectly using scanPlanSQLScanner instead of
scanPlanCodecSQLScanner, which resulted in the source type being a
string.